Moved font and color menus into components.
- Added ColorMenu component
- Added FontMenu component
- Moved associated elements into the respective component
