+++ src/components/index.js
@@ -5,3 +5,5 @@ export { default as ListComp } from './ListComp';
 export { default as NavLink } from './NavLink';
 export { default as Header } from './Header';
 export { default as Footer } from './Footer';
+export { default as FontMenu } from './FontMenu';
+export { default as ColorMenu } from './ColorMenu';
