+++ src/containers/App.js
@@ -4,7 +4,7 @@ import { IndexLink } from 'react-router';
 import { NavLink } from '../components';
 
 class App extends Component {
-  render () {
+  render() {
     return (
       <div>
         <div
